Add gtest support

+++ src/client/ChatClient.hpp
@@ -1,17 +1,18 @@
 #pragma once
 #include <string>
+#include <cstdint>
 
 namespace client
 {
     class ChatClient
     {
     public:
-        ChatClient(const std::string serverIp, int port);
+        ChatClient(const std::string serverIp, uint16_t port);
         void run();
 
     private:
         std::string serverIp;
-        int port;
+        uint16_t port;
         int sock;
 
         void receiveMessages();
